Improve bundling speed replacing inefficient code
Current bundling function `filterFiles` reduces an array of file names to an object `filesMap` having those file names as properties and `true` as value to later track keep/remove status.

The problem is that the reduce operation creates a new object on each iteration and when the list of files is large enough, it stresses Node.js' memory usage and garbage collection processes to a point where it becomes very inneficient.

Instead, a much more efficient way is to just create an empty object and mutate it by adding each property as required.

For a 20,000+ files bundle, the new approach is aprox 20x faster.